First build of data cleaning backed by a lot of exploration
Added a number of descriptive characteristics to the dataframe (to indicate incomplete information, or an item issue).

The "Description" field has been tidied up to identify where there were issues with the order that were recorded in the "Description" column rather than as a seperate field.

Data now in a state to start performing more analytics (on sliced data) for some segmentation analysis, RFM, etc.